extract worklog
